app/requirements.txt
# sqlalchemy
python-dotenv